requirements.txt
huggingface_hub

llama-cpp-python
llama_cpp_agent

ollama

langchain-ollama
langchain
langchain_community
